Added some horizontal movement controls

+++ include/circle.h
@@ -3,7 +3,7 @@
 
 class Circle : public Shape {
 public:
-    Circle(Vector center, int radius, bool isRigidBody=false);
+    Circle(Vector center, float radius, bool isRigidBody=false);
     ~Circle();
     void draw(SDL_Renderer *renderer) override;
